Published latest changes to the sample

+++ entities/Link.java
@@ -3,12 +3,11 @@ package entities;
 import java.io.Serializable;
 import java.util.Date;
 
-
 @SuppressWarnings("serial")
 public class Link implements Serializable {
 	public String Token;
 
-	public int SyncPointId;
+	public long SyncPointId;
 
 	public String VirtualPath;
